requirements.txt
fastapi
uvicorn
pytesseract
pillow
pdf2image
pydantic